+++ beacon/sync/src/test/java/tech/pegasys/teku/beacon/sync/forward/multipeer/batches/SyncSourceBatchTest.java
@@ -44,7 +44,7 @@ import tech.pegasys.teku.networking.eth2.rpc.beaconchain.methods.BlocksByRangeRe
 import tech.pegasys.teku.networking.p2p.peer.PeerDisconnectedException;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecarOld;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.versions.deneb.BeaconBlockBodyDeneb;
@@ -146,7 +146,7 @@ public class SyncSourceBatchTest {
 
     // only receiving last block (70 + 50 - 1)
     final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock(119);
-    final List<BlobSidecar> blobSidecars = dataStructureUtil.randomBlobSidecarsForBlock(block);
+    final List<BlobSidecarOld> blobSidecars = dataStructureUtil.randomBlobSidecarsForBlock(block);
 
     receiveBlocks(batch, block);
     receiveBlobSidecars(batch, blobSidecars);
@@ -240,7 +240,7 @@ public class SyncSourceBatchTest {
 
     final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock(19);
 
-    final List<BlobSidecar> blobSidecars =
+    final List<BlobSidecarOld> blobSidecars =
         new ArrayList<>(dataStructureUtil.randomBlobSidecarsForBlock(block));
     // receiving more sidecars than expected
     blobSidecars.add(
@@ -266,7 +266,7 @@ public class SyncSourceBatchTest {
     final int numberOfKzgCommitments =
         BeaconBlockBodyDeneb.required(block.getMessage().getBody()).getBlobKzgCommitments().size();
     // receiving sidecars with different slot than the block
-    final List<BlobSidecar> blobSidecars =
+    final List<BlobSidecarOld> blobSidecars =
         IntStream.range(0, numberOfKzgCommitments)
             .mapToObj(
                 index ->
@@ -294,8 +294,8 @@ public class SyncSourceBatchTest {
 
     final SignedBeaconBlock block = dataStructureUtil.randomSignedBeaconBlock(19);
 
-    final List<BlobSidecar> blobSidecars = dataStructureUtil.randomBlobSidecarsForBlock(block);
-    final List<BlobSidecar> unexpectedBlobSidecars = new ArrayList<>(blobSidecars);
+    final List<BlobSidecarOld> blobSidecars = dataStructureUtil.randomBlobSidecarsForBlock(block);
+    final List<BlobSidecarOld> unexpectedBlobSidecars = new ArrayList<>(blobSidecars);
     // receiving sidecars with unknown roots
     unexpectedBlobSidecars.addAll(
         dataStructureUtil.randomBlobSidecarsForBlock(
@@ -363,8 +363,8 @@ public class SyncSourceBatchTest {
     getSyncSource(batch).receiveBlocks(blocks);
   }
 
-  protected void receiveBlobSidecars(final Batch batch, final List<BlobSidecar> blobSidecars) {
-    getSyncSource(batch).receiveBlobSidecars(blobSidecars.toArray(new BlobSidecar[] {}));
+  protected void receiveBlobSidecars(final Batch batch, final List<BlobSidecarOld> blobSidecars) {
+    getSyncSource(batch).receiveBlobSidecars(blobSidecars.toArray(new BlobSidecarOld[] {}));
   }
 
   protected void requestError(final Batch batch, final Throwable error) {