+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/BlockContainer.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 import tech.pegasys.teku.infrastructure.ssz.SszContainer;
 import tech.pegasys.teku.infrastructure.ssz.SszData;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecarOld;
 import tech.pegasys.teku.spec.datastructures.blocks.versions.deneb.BlockContents;
 
 /**
@@ -34,7 +34,7 @@ public interface BlockContainer extends SszData, SszContainer {
     return getBlock().getSlot();
   }
 
-  default Optional<List<BlobSidecar>> getBlobSidecars() {
+  default Optional<List<BlobSidecarOld>> getBlobSidecars() {
     return Optional.empty();
   }