+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BlobSidecarsByRangeListenerValidatingProxyTest.java
@@ -30,7 +30,7 @@ import tech.pegasys.teku.networking.eth2.peers.Eth2Peer;
 import tech.pegasys.teku.networking.p2p.rpc.RpcResponseListener;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecarOld;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
 public class BlobSidecarsByRangeListenerValidatingProxyTest {
@@ -42,7 +42,7 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
   private final KZG kzg = mock(KZG.class);
 
   @SuppressWarnings("unchecked")
-  private final RpcResponseListener<BlobSidecar> listener = mock(RpcResponseListener.class);
+  private final RpcResponseListener<BlobSidecarOld> listener = mock(RpcResponseListener.class);
 
   @BeforeEach
   void setUp() {
@@ -60,7 +60,7 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
 
     final SafeFuture<?> result = listenerWrapper.onResponse(blobSidecar1);
@@ -82,7 +82,7 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
         new BlobSidecarsByRangeListenerValidatingProxy(
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
-    final BlobSidecar blobSidecar0 =
+    final BlobSidecarOld blobSidecar0 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ZERO, dataStructureUtil.randomBytes32(), UInt64.ZERO);
 
@@ -106,18 +106,18 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar10 =
+    final BlobSidecarOld blobSidecar10 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
-    final BlobSidecar blobSidecar11 =
+    final BlobSidecarOld blobSidecar11 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ONE);
     final Bytes32 blockRoot2 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(2), blockRoot2, blockRoot1, UInt64.ZERO);
     final Bytes32 blockRoot3 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar3 =
+    final BlobSidecarOld blobSidecar3 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(3), blockRoot3, blockRoot2, UInt64.ZERO);
     final Bytes32 blockRoot4 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar4 =
+    final BlobSidecarOld blobSidecar4 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(4), blockRoot4, blockRoot3, UInt64.ZERO);
 
     assertDoesNotThrow(() -> listenerWrapper.onResponse(blobSidecar10).join());
@@ -137,19 +137,19 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
     final Bytes32 blockRoot2 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(3), blockRoot2, blockRoot1, UInt64.ZERO);
     final Bytes32 blockRoot3 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar3 =
+    final BlobSidecarOld blobSidecar3 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(5), blockRoot3, blockRoot2, UInt64.ZERO);
     final Bytes32 blockRoot4 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar4 =
+    final BlobSidecarOld blobSidecar4 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(8), blockRoot4, blockRoot3, UInt64.ZERO);
     final Bytes32 blockRoot5 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar5 =
+    final BlobSidecarOld blobSidecar5 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(9), blockRoot5, blockRoot4, UInt64.ZERO);
 
     listenerWrapper.onResponse(blobSidecar1).join();
@@ -177,10 +177,10 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
     final Bytes32 blockRoot2 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.valueOf(2), blockRoot2, dataStructureUtil.randomBytes32(), UInt64.ZERO);
 
@@ -207,23 +207,23 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ONE);
-    final BlobSidecar blobSidecar3 =
+    final BlobSidecarOld blobSidecar3 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(2));
-    final BlobSidecar blobSidecar4 =
+    final BlobSidecarOld blobSidecar4 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(3));
-    final BlobSidecar blobSidecar5 =
+    final BlobSidecarOld blobSidecar5 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(4));
-    final BlobSidecar blobSidecar6 =
+    final BlobSidecarOld blobSidecar6 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(5));
-    final BlobSidecar blobSidecar7 =
+    final BlobSidecarOld blobSidecar7 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(6));
 
@@ -255,11 +255,11 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ONE);
-    final BlobSidecar blobSidecar3 =
+    final BlobSidecarOld blobSidecar3 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(3));
 
@@ -286,7 +286,7 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, UInt64.ONE);
 
     final SafeFuture<?> result = listenerWrapper.onResponse(blobSidecar1);
@@ -309,10 +309,10 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
     final Bytes32 blockRoot2 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.valueOf(2), blockRoot2, blockRoot1, UInt64.ONE);
 
     assertDoesNotThrow(() -> listenerWrapper.onResponse(blobSidecar1).join());
@@ -338,11 +338,11 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ONE);
-    final BlobSidecar blobSidecar4 =
+    final BlobSidecarOld blobSidecar4 =
         dataStructureUtil.randomBlobSidecar(
             UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.valueOf(3));
 
@@ -369,10 +369,10 @@ public class BlobSidecarsByRangeListenerValidatingProxyTest {
             spec, peer, listener, maxBlobsPerBlock, kzg, startSlot, count);
 
     final Bytes32 blockRoot1 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar1 =
+    final BlobSidecarOld blobSidecar1 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot1, Bytes32.ZERO, UInt64.ZERO);
     final Bytes32 blockRoot2 = dataStructureUtil.randomBytes32();
-    final BlobSidecar blobSidecar2 =
+    final BlobSidecarOld blobSidecar2 =
         dataStructureUtil.randomBlobSidecar(UInt64.ONE, blockRoot2, blockRoot1, UInt64.ZERO);
 
     assertDoesNotThrow(() -> listenerWrapper.onResponse(blobSidecar1).join());