+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BlobSidecarsByRootMessageHandlerTest.java
@@ -48,7 +48,7 @@ import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.config.SpecConfigDeneb;
-import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecar;
+import tech.pegasys.teku.spec.datastructures.blobs.versions.deneb.BlobSidecarOld;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobIdentifier;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessage;
 import tech.pegasys.teku.spec.datastructures.networking.libp2p.rpc.BlobSidecarsByRootRequestMessageSchema;
@@ -78,14 +78,14 @@ public class BlobSidecarsByRootMessageHandlerTest {
 
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil(spec);
 
-  private final ArgumentCaptor<BlobSidecar> blobSidecarCaptor =
-      ArgumentCaptor.forClass(BlobSidecar.class);
+  private final ArgumentCaptor<BlobSidecarOld> blobSidecarCaptor =
+      ArgumentCaptor.forClass(BlobSidecarOld.class);
 
   private final ArgumentCaptor<RpcException> rpcExceptionCaptor =
       ArgumentCaptor.forClass(RpcException.class);
 
   @SuppressWarnings("unchecked")
-  private final ResponseCallback<BlobSidecar> callback = mock(ResponseCallback.class);
+  private final ResponseCallback<BlobSidecarOld> callback = mock(ResponseCallback.class);
 
   private final CombinedChainDataClient combinedChainDataClient =
       mock(CombinedChainDataClient.class);
@@ -221,7 +221,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
 
     final List<Bytes32> respondedBlobSidecarBlockRoots =
         blobSidecarCaptor.getAllValues().stream()
-            .map(BlobSidecar::getBlockRoot)
+            .map(BlobSidecarOld::getBlockRoot)
             .collect(Collectors.toUnmodifiableList());
     final List<Bytes32> blobIdentifiersBlockRoots =
         List.of(
@@ -281,7 +281,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
 
     // let first blobSidecar slot will be earlier than the minimum_request_epoch (for this test it
     // is denebForkEpoch)
-    final BlobSidecar blobSidecar = mock(BlobSidecar.class);
+    final BlobSidecarOld blobSidecar = mock(BlobSidecarOld.class);
     when(blobSidecar.getSlot()).thenReturn(UInt64.ONE);
     when(combinedChainDataClient.getBlobSidecarByBlockRootAndIndex(
             blobIdentifiers.get(0).getBlockRoot(), blobIdentifiers.get(0).getIndex()))
@@ -323,7 +323,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
 
     verify(callback, times(5)).respond(blobSidecarCaptor.capture());
 
-    final List<BlobSidecar> sentBlobSidecars = blobSidecarCaptor.getAllValues();
+    final List<BlobSidecarOld> sentBlobSidecars = blobSidecarCaptor.getAllValues();
 
     // Requesting 5 blob sidecars
     verify(peer, times(1)).approveBlobSidecarsRequest(any(), eq(Long.valueOf(5)));
@@ -335,7 +335,7 @@ public class BlobSidecarsByRootMessageHandlerTest {
         .forEach(
             index -> {
               final BlobIdentifier identifier = blobIdentifiers.get(index);
-              final BlobSidecar blobSidecar = sentBlobSidecars.get(index);
+              final BlobSidecarOld blobSidecar = sentBlobSidecars.get(index);
               assertThat(blobSidecar.getBlockRoot()).isEqualTo(identifier.getBlockRoot());
               assertThat(blobSidecar.getIndex()).isEqualTo(identifier.getIndex());
             });